requirements.txt
dashscope>=1.23.0
requests>=2.32.0
opencv-python>=4.9.0
numpy>=1.24.0
torch>=2.1.0
Pillow>=9.0.0
